ci: agregar GitHub Actions, suite de tests unitarios y limpiar código
Tests (16 passing):
- tests/conftest.py: fixtures session-scoped con StaticPool y force-login
  vía session_transaction (yield fuera del app_context para que cada
  request tenga su propio g y Flask-Login no cachee AnonymousUser)
- tests/test_autorizacion.py: 11 tests de autorización (anónimo, admin,
  créditos insuficientes, sesión MV)
- tests/test_pagos.py: 5 tests de pagos (idempotencia webhook, firma HMAC,
  orden inválida)
- tests/integration/: scripts manuales reubicados, ignorados en CI

CI (.github/workflows/ci.yml):
- Python 3.11, flake8 sobre módulos core, pytest --ignore=tests/integration

Limpieza de código:
- Renombrar modulos/modelos.py → modulos/dataclasses_expediente.py
  (evitar ambigüedad con models.py)
- Float → Numeric(10,2) para monto_pagado (precisión monetaria)
- User.query.get() → db.session.get() (SQLAlchemy 2.0)
- Eliminar requirements-fase*.txt obsoletos
- Corregir imports sin uso en extensions.py, admin.py, contacto.py,
  descargas.py, jurisprudencia.py, auth.py, pagos.py

+++ modulos/models.py
@@ -23,7 +23,7 @@ Uso:
 
 import os
 import json
-from datetime import datetime, timedelta
+from datetime import datetime
 from werkzeug.security import generate_password_hash, check_password_hash
 from cryptography.fernet import Fernet, InvalidToken
 from flask_login import UserMixin
@@ -267,7 +267,7 @@ class CompraCreditos(db.Model):
 
     # Detalles de la compra
     creditos_comprados = db.Column(db.Integer, nullable=False)
-    monto_pagado = db.Column(db.Float, nullable=False)  # En USD/ARS
+    monto_pagado = db.Column(db.Numeric(10, 2), nullable=False)
     plan = db.Column(db.String(50), nullable=False)  # free, pro, premium
 
     # Estado
